rename DB files name
